Limit the # of cpus each Java service can see with -XX:ActiveProcessorCount=N
A very very initial and desperate way to limit the number of cpus
available to each java service so starting up the docker composition
does not bog down the machine.

REVISIT: this should actually be controlled with cgroups (which afaik
docker-compose does not support), hence at the very least it could be
done with docker swarm, setting the limist in the docker-compose.yml
file.