Move config/ to the project's root dir and copy it to config-service's classpath
Have the `config/` directory at the project root directory, create a
bind mount to `config-service/opt/app/config` in `docker-compose.yml`.

The contents of `config/` are copied to the config service's classpath
at build time, though it's recommended to use an external source (either
file or git). The `native` active profile and the default location of
`./config` (with the docker container's working dir being `/opt/app`)
makes for an easy way to follow the defaults.