Fix FilteringXmlBeanDefinitionReader and cleaner semantics
Receives a single regular expression that represents which beans to
include, not to exclude. Added clear examples for including and
excluding based on bean names, and fixed the regular expressions used to
load the gs-main applicationContext.xml excluding the beans that are
in charge of implementations of `GeoServerBackendConfigurer`, and the
wms-service's includes of gs-wfs beans.

Fixes wms-service, it couldn't respond to GetMap queries before due to
missing wfs related beans.