Do not ser servletcontext to GeoSeverResourceLoader
Thhe configurer set is in charge of setting the ResourceLoader's
base directory. Setting the servlet context would lead to it trying to
figure out the data directory location and end up setting it as Tomcat's
temporary folder. Better make the application fail if the configuration
does not set a valid location. Some times it's just a local cache
directory, depends on each backend implementation.