Support and configure default resource limits in docker composition
Add CPU and memory limits to the service definitions in
docker-compose.yml.

Need to run docker-compose with the --compatibility flag in order for
deploy.resources.limits.* to be considered (otherwise they're only
obeyed in a swarm deploy).

Change JAVA_OPTS to use -XX:MaxRAMPercentage=[0..100]
-XX:MaxRAM=<size><M|G> and hence set the max heap size dynamically based
on the available RAM.

The --MaxRAM is set in case docker-compose wasn't (accidentally?) run
with --compatibility. Otherwise it's unnecessary given the percentage
set and the memory limit imposed to the container.

Remove command overrides to use dockerize in all services except config
to wait for discovery, and let the retry configuration be used instead.

Add a log entry when each service is ready to verify the memory and cpu
limits are seen by the JVM.