Switch to curl based healthcheck command in `Dockerfile`s
Running a custom java app for healthcheck can be too
resource-consuming due to JVM startup time.

The current base image does come with `curl` installed,
so at least for the time being use that.

It may be that we use a custom healthckeck back in the
future so we don't depend on any external command, but
for the time being let's use curl.