gwc: fix blobstore remote event handling, CompositeBlobStore didn't notice
Upstream's `CompositeBlobStore` registers itself as a
`BlobStoreConfigurationListener`.

When the `RemoteBlobstoreEvent`s came in through the event
bus, the xml configuration was correctly reloaded, but
`CompositeBlobStore`, which holds the `BlobStoreInfo`s
as internal state, didn't notice.

Now `CloudGwcXmlConfiguration` will notify its (legacy style)
listeners when `BlobStoreEvent` is being handled, as a
result of `GeoWebCacheRemoteEventsBroker` publishing
a conveying a remote event as a local event.