Fix invalid virtual service endpoints not returning 404
If an invalid virtual service request is made (e.g.
to a non existing workspace or layer), upstream
GeoServer will return a 404 thanks to `OWSHandlerMapping`
checking for the workspace/layergroup/layer and returning
a `null` request handler.

Spring-boot does not automatically engages the `OWSHandlerMapping`
beans in the request lifecycle.

This patch creates separate controller entry points for
`/{virtualService}/<ows service>` and
`/{virtualService}/{layer}/<ows service>`, using a helper
service bean to verify the existence of the requested
workspace/layergroup and layer, returning a 404 with
an explanatory message if the virtual service is not found,
and hence avoiding to return the full layer list in
getcapabilities requests instead.