Do send no-change remote events, they still increment the update sequence
A typical no-change event is when through the web-ui a Layer name
is changed. That's gonna send two events, one for the underlying
resource with the actual name change (e.g. the FeatureTypeInfo),
and one for the Layer, which didn't actually change its name since
it's pegged to its resource's.