Disable caching for JDBCDirectoryStructure
JDBCDirectoryStructure gets its cache only from
DefaultCacheProvider.findProvider().

Unfortunately there's no sane way of disabling it other than setting the
default provider to no-op cache.

And it is necessary because the resources in the cache will never be up
to date with the database contents, which prevents starting up from an
empty database, since all instances will think the geoserver security
manager resources are uninitialized and hence try to upgrade the
configuration;