Change GeoTools log redirection from Log4j2 to Logback
GeoTools version of Log4j2 (2.17.2) and the one used by spring-boot
(2.21.1) differ in the  API used by GeoTools to set the logging level,
which produces an exception due to a missing setLevel(...) method in the
new version.

Change log redirection to use Logback, and make sure the
ApplicationContextInitializers run only once, as they may be called
twice, once for the webapp context, and again for the Actuator context.