qa: Remove all superfluously overridden methods from CatalogPlugin
`CatalogPlugin` started by only implementing `Catalog` and hence
duplicated most of upstream's `CatalogImpl` functionality. Then it was
required for it to extend `CatalogImpl` instead, due to code in
GeoServer assuming the `Catalog` **is A** `CatalogImpl`, but the
overridden methods that did the same weren't removed.

This leaves `CatalogPlugin` much cleaner and it's easier to see how it
exactly differ from its superclass.