Remove unused catalog-service microservice and client
The catalog service has been out of commission for a long time now,
becoming dead code.

The idea is still sound but the complexity of WebFlux makes it a
maintenance nightmare. Plus, with Java 21's virtual threads, it will be
possible to write a non WebFlux implementation with almost the same
benefits, except maybe for the reactive back-pressure support.