PgsqlResourceStore performance improvement
Change the resourcestore table to hold the full path instead of just the name,
so no recursive queries are required to query. And delete the resources view.
The caveat being the ResourceStore implementation is hence in charge of
transactionally update the path of all children when a directory resource is
renamed/moved, but the perfomance gain is huge.