share webui authentication with the other services through the gateway
Introduce a `GatewaySharedAuthenticationFilter`, where the web-ui
service acts as service, returning the authenticated user name and roles
as response headers, the other services act as client, receiving the
user name and roles as request headers, and the gateway acts as
intermediary to convey the auth returned by the webui and sending it to
the other services.

The functionality is enabled by default, and automatically creates the
required authentication filter and injects it to the filter chains.

This behavior can be controlled through the following configuration
properties:

```
  geoserver:
    security:
      gateway-shared-auth:
         enabled: true
         auto: true
         server: false (true for webui)
```