Fix RepositoryCatalogFacadeImpl's query of PublishedInfo.class
The webui layer preview page, and possibly other code paths, lead to
querying the `PublishedInfo` super interface for `LayerInfo` and
`LayerGroupInfo`.

These queries must be delegated to the layer and layer group queries.
The `DefaultCatalogFacade` in GeoServer does it and builds an in-memory
list of the results.

Our `RepositoryCatalogFacadeImpl` does it in a streaming fashion, but it
wasn't respecting the limits, resulting in wrong results for the layer
preview page.

Now it queries `LayerRepository` and `LayerGroupRepository`, and returns
a merge-sorted  stream.

If the query does not specify a sort order, the `id` property is used to provide
predictable order for the merge-sort algorithm.

When the returned stream is closed, it'll close the two underlying streams.