Avoid exceptions when gateway-shared-auth is disabled after being enabled
If the `gateway-shared-auth` geoserver authentication filter was
enabled, and it was automatically added to the filter chains, when
disabled and the applications restart, would produce an error message
and the webui Authentication configuration page would be broken.

This patch adds a no-op filter when `gateway-shared-auth` is disabled by
externalized configuration.