Fix for GWC not storing tiles when using datadir or jdbcconfig catalog back-ends
Abstract out the PgconfigGwcInitializer as AbstractGwcInitializer and
refactor the default configuration classes to include the initializer.

Tiles weren't being stored when using the `datadir` or `jdbcconfig`
catalog backends because `ConfigurableBlobStore.setChanged()` wouldn't
be called at startup and it would hence bypass the storage, as its
internal `configured` flag was not set.