Fix saving workspace services not working with pgconfig
Saving the workspace services not working with pgconfig when
caching is enabled. `CachingGeoServerFacade` assumed there's
nothing to do when a `ServiceInfo` is added for a workspace,
but it was caching the `null` value, which is expected to save
db roundtrips.

- Add missing treatment of `ServiceInfoAdded` and `SettingInfoAdded`
to `CachingGeoServerFacade`.

- Add referential integrity for settingsinfo and serviceinfo

Now the flyway migration will add referential integrity for:

 - `settingsinfo(workspace) -> workspaceindo(id)`
 - `serviceinfo(workspace) -> workspaceinfo(id)`

And unique indexes for:

- `settingsinfo(workspace)`
- `serviceinfo("@type", workspace)`

As a side effect of the caching facade returning null once a
`ServiceInfo` as added, you could add multiple services of the
same kind for a workspace. The referential integrity checks will
avoid that, and the FlyWay migration will delete existing duplicates.